# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Aios Practice Platform API
  version: 1.0.0
  description: 题练通一期正式 OpenAPI 基线。
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
...
paths:
...
  /users:
    get:
      tags: [User]
      summary: 用户列表
      parameters:
        - in: query
          name: user_type
          schema: { type: string }
        - in: query
          name: keyword
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: page_size
          schema: { type: integer }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
    post:
      tags: [User]
      summary: 创建用户
      requestBody:
        $ref: '#/components/requestBodies/UserBody'
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /users/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [User]
      summary: 用户详情
      responses:
        '200': { $ref: '#/components/responses/Ok' }
    put:
      tags: [User]
      summary: 更新用户
      requestBody:
        $ref: '#/components/requestBodies/UserBody'
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /users/{id}/roles:
    parameters:
      - $ref: '#/components/parameters/Id'
    put:
      tags: [User]
      summary: 给用户分配角色
      requestBody:
        $ref: '#/components/requestBodies/UserRolesBody'
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /users/{id}/reset-password:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      tags: [User]
      summary: 重置密码
      requestBody:
        $ref: '#/components/requestBodies/ResetPasswordBody'
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /users/{id}/disable:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      tags: [User]
      summary: 禁用用户
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /student-transitions:
    get:
      tags: [Audit]
      summary: 学籍变更列表
      parameters:
        - in: query
          name: student_id
          schema:
            type: integer
            format: int64
        - in: query
          name: transition_type
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200': { $ref: '#/components/responses/StudentTransitionPageOk' }
    post:
      tags: [Audit]
      summary: 记录学籍变更
      requestBody:
        $ref: '#/components/requestBodies/StudentTransitionBody'
      responses:
        '200': { $ref: '#/components/responses/StudentTransitionOk' }
  /teacher-assignment-histories:
    get:
      tags: [Audit]
      summary: 任课变更历史列表
      parameters:
        - in: query
          name: teacher_id
          schema:
            type: integer
            format: int64
        - in: query
          name: class_id
          schema:
            type: integer
            format: int64
        - in: query
          name: course_id
          schema:
            type: integer
            format: int64
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200': { $ref: '#/components/responses/TeacherAssignmentHistoryPageOk' }
  /teacher-assignment-changes:
    post:
      tags: [Audit]
      summary: 记录任课变更
      requestBody:
        $ref: '#/components/requestBodies/TeacherAssignmentChangeBody'
      responses:
        '200': { $ref: '#/components/responses/TeacherAssignmentHistoryOk' }
...
  /audit-logs:
    get:
      tags: [Audit]
      summary: 审计日志列表
      operationId: listAuditLogs
      parameters:
        - in: query
          name: module_name
          schema:
            type: string
        - in: query
          name: resource_type
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200': { $ref: '#/components/responses/AuditLogPageOk' }
  /entity-snapshots:
    get:
      tags: [Audit]
      summary: 实体快照列表
      operationId: listEntitySnapshots
      parameters:
        - in: query
          name: entity_type
          schema:
            type: string
        - in: query
          name: entity_id
          schema:
            type: integer
            format: int64
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200': { $ref: '#/components/responses/EntitySnapshotPageOk' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
  responses:
    Ok:
      description: OK
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BaseResponse'
...
    AuditLogPageOk:
      description: 审计日志分页列表
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuditLogPageResponse'
    EntitySnapshotPageOk:
      description: 实体快照分页列表
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/EntitySnapshotPageResponse'
    StudentTransitionOk:
      description: 学籍变更详情
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StudentTransitionResponse'
    StudentTransitionPageOk:
      description: 学籍变更分页列表
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StudentTransitionPageResponse'
    TeacherAssignmentHistoryOk:
      description: 任课变更详情
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TeacherAssignmentHistoryResponse'
    TeacherAssignmentHistoryPageOk:
      description: 任课变更分页列表
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TeacherAssignmentHistoryPageResponse'
...
  requestBodies:
...
    UserBody:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UserInput'
    UserRolesBody:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UserRolesInput'
    ResetPasswordBody:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResetPasswordInput'
...
    StudentTransitionBody:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StudentTransitionInput'
    TeacherAssignmentChangeBody:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TeacherAssignmentChangeInput'
...
  schemas:
    BaseResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
        data:
          nullable: true
        request_id:
          type: string
...
    UserInput:
      type: object
      required: [username, display_name, user_type]
      properties:
        username:
          type: string
        display_name:
          type: string
        user_type:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
        role_ids:
          type: array
          items:
            type: integer
            format: int64
    UserRolesInput:
      type: object
      required: [role_ids]
      properties:
        role_ids:
          type: array
          items:
            type: integer
            format: int64
    ResetPasswordInput:
      type: object
      required: [new_password]
      properties:
        new_password:
          type: string
...
    AuditLogItem:
      type: object
      required: [id, tenant_id, module_name, action_name, resource_type, result, created_at]
      properties:
        id:
          type: integer
          format: int64
        tenant_id:
          type: integer
          format: int64
        operator_user_id:
          type: integer
          format: int64
          nullable: true
        module_name:
          type: string
        action_name:
          type: string
        resource_type:
          type: string
        resource_id:
          type: integer
          format: int64
          nullable: true
        before_json:
          type: object
          additionalProperties: true
          nullable: true
        after_json:
          type: object
          additionalProperties: true
          nullable: true
        request_id:
          type: string
          nullable: true
        ip:
          type: string
          nullable: true
        user_agent:
          type: string
          nullable: true
        result:
          type: string
        created_at:
          type: string
          format: date-time
    EntitySnapshotItem:
      type: object
      required: [id, tenant_id, entity_type, entity_id, snapshot_type, snapshot_json, version_no, created_at]
      properties:
        id:
          type: integer
          format: int64
        tenant_id:
          type: integer
          format: int64
        entity_type:
          type: string
        entity_id:
          type: integer
          format: int64
        snapshot_type:
          type: string
        snapshot_json:
          type: object
          additionalProperties: true
        version_no:
          type: integer
        trigger_event_type:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    StudentTransitionInput:
      type: object
      required: [student_id, transition_type, occurred_at]
      properties:
        student_id:
          type: integer
          format: int64
        transition_type:
          type: string
          enum: [class_change, promote, transfer_in, transfer_out, graduate, leave_school, re_enroll]
        to_class_id:
          type: integer
          format: int64
          nullable: true
        occurred_at:
          type: string
          format: date-time
        remark:
          type: string
          nullable: true
    StudentTransitionItem:
      type: object
      required: [id, tenant_id, student_id, transition_type, occurred_at, operator_id, created_at]
      properties:
        id:
          type: integer
          format: int64
        tenant_id:
          type: integer
          format: int64
        student_id:
          type: integer
          format: int64
        transition_type:
          type: string
        from_school_id:
          type: integer
          format: int64
          nullable: true
        from_grade_id:
          type: integer
          format: int64
          nullable: true
        from_class_id:
          type: integer
          format: int64
          nullable: true
        to_school_id:
          type: integer
          format: int64
          nullable: true
        to_grade_id:
          type: integer
          format: int64
          nullable: true
        to_class_id:
          type: integer
          format: int64
          nullable: true
        occurred_at:
          type: string
          format: date-time
        operator_id:
          type: integer
          format: int64
        remark:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    TeacherAssignmentChangeInput:
      type: object
      required: [teacher_id, class_id, course_id, change_type, effective_at]
      properties:
        teacher_id:
          type: integer
          format: int64
        class_id:
          type: integer
          format: int64
        course_id:
          type: integer
          format: int64
        change_type:
          type: string
          enum: [assign, unassign]
        effective_at:
          type: string
          format: date-time
    TeacherAssignmentHistoryItem:
      type: object
      required: [id, tenant_id, teacher_id, class_id, course_id, change_type, effective_from, operator_id, created_at]
      properties:
        id:
          type: integer
          format: int64
        tenant_id:
          type: integer
          format: int64
        teacher_id:
          type: integer
          format: int64
        class_id:
          type: integer
          format: int64
        course_id:
          type: integer
          format: int64
        change_type:
          type: string
        effective_from:
          type: string
          format: date-time
        effective_to:
          type: string
          format: date-time
          nullable: true
        operator_id:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
    AuditLogPageResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/AuditLogItem'
            page:
              type: integer
            page_size:
              type: integer
            total:
              type: integer
        request_id:
          type: string
    EntitySnapshotPageResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/EntitySnapshotItem'
            page:
              type: integer
            page_size:
              type: integer
            total:
              type: integer
        request_id:
          type: string
    StudentTransitionResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/StudentTransitionItem'
        request_id:
          type: string
    StudentTransitionPageResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/StudentTransitionItem'
            page:
              type: integer
            page_size:
              type: integer
            total:
              type: integer
        request_id:
          type: string
    TeacherAssignmentHistoryResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/TeacherAssignmentHistoryItem'
        request_id:
          type: string
    TeacherAssignmentHistoryPageResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/TeacherAssignmentHistoryItem'
            page:
              type: integer
            page_size:
              type: integer
            total:
              type: integer
        request_id:
          type: string
...